TMItransient

Documentation for TMItransient.

TMItransient.EvolvingField

TMItransient.agedistribution

TMItransient.ces_ncwrite

TMItransient.deltaresponse

TMItransient.globalmean_impulseresponse

TMItransient.goodtime

TMItransient.read_stepresponse

TMItransient.readopt

TMItransient.s_array

TMItransient.setupODE_nojac

TMItransient.stability_check

TMItransient.stepresponse

TMItransient.taudeltaresponse

TMItransient.varying!

TMItransient.vintagedistribution

TMItransient.EvolvingField — Type

struct EvolvingField

This structure permits the grid to be 
automatically passed to functions with
the evolving tracer field.

This structure assumes the Tracer type to be 
three-dimensional with an additional vector dimension for time.

tracer::Vector{Array{T,3}}
γ::Grid
name::Symbol
longname::String
units::String

TMItransient.agedistribution — Method

function agedistribution

age distribution refers to distribution over lags, not space
sometimes called a transit time distribution

TMItransient.ces_ncwrite — Method

function ces_ncwrite(γ,time,sol_array)
Write .nc file output for commonerasim.jl

Arguments

γ:

time: vector of time values

sol_array: solution array in form time x lat x lon x depth - must match γ + time

Output

saves .nc file titled "ces_output.nc" in data array

TMItransient.deltaresponse — Method

function deltaresponse

Take CDF and turn it into PDF

TMItransient.globalmean_impulseresponse — Method

function globalmean_impulseresponse

Ḡ: satisfied ∫₀^∞ Ḡ(τ) dτ = 1

`alg`: centered or leapfrog

Does leapfrog satisfy normalization?

TMItransient.read_stepresponse — Method

function read_stepresponse()

read previously computed MATLAB output using shell script

Warning: hard-coded file names from a 4° x 4° TMI run

TMItransient.readopt — Method

 read surface layer

TMItransient.s_array — Method

function  s_array(sol, γ)
Converts from DE.jl output to time x lat x lon x depth

TMItransient.setupODE_nojac — Method

function setupODE(γ, u0,tsfc, dsfc,bc,L, t_int)
Setup ODEFunction for LC+Bf with no Jacobian

TMItransient.stability_check — Method

stability_check(sol_array, Csfc) 
checks stability of ODE output

Arguments

sol_array: solution array

Csfc: boundary condition

Output

prints true or false

TMItransient.taudeltaresponse — Method

function taudeltaresponse

Take CDF and turn it into PDF, get lag timescale

Seems inefficient, reading file for time lag alone

TMItransient.varying! — Method

function varying!(du, u, p, t)
ODE function for varying boundary cond
Sets up dc/dt = L*C + B*f to be solved

Arguments

du: dc/dt (must have this name for DifferentialEquations.jl to work

u: C, what we are solving for

p: parameters for diffeq - must hold specified vars

t: time we are solving for (automatically determined by DE.jl)

Output

du: numerical value of LC+Bf, vector of size 74064 for 4°

TMItransient.vintagedistribution — Method

function vintagedistribution(t₀,tf,Δ,τ,tmodern=2022,interp="linear")

percentage of water in the modern ocean
from a vintage defined by the calendar year interval
t₀ [cal yr CE] => tf [cal yr CE]

Arguments

t₀: Starting calendar year of vintage, e.g., 1450 CE

tf: final calendar year of vintage, e.g., 1850 CE

Δ::Vector{Field}: Step function response

τ = Vector{Float64}: time lags associated with step response

tmodern=2022: modern calendar year

interp=linear: or can be "spline"

Output

g::Field: 3D distribution of vintage contribution

Warning

should be a way to make Δ argument more general (more types)

TMItransient.goodtime — Function

function gooddata
a useful one-liner

TMItransient.stepresponse — Method

function stepresponse(loc)

response to a Heaviside function at a `loc`