TMItransient

Documentation for TMItransient.

TMItransient.ces_ncwrite

TMItransient.readopt

TMItransient.s_array

TMItransient.setupODE_nojac

TMItransient.stability_check

TMItransient.varying!

TMItransient.ces_ncwrite — Method

function ces_ncwrite(γ,time,sol_array)
Write .nc file output for commonerasim.jl

Arguments

γ:

time: vector of time values

sol_array: solution array in form time x lat x lon x depth - must match γ + time

Output

saves .nc file titled "ces_output.nc" in data array

TMItransient.readopt — Method

 read surface layer

TMItransient.s_array — Method

function  s_array(sol, γ)
Converts from DE.jl output to time x lat x lon x depth

TMItransient.setupODE_nojac — Method

function setupODE(γ, u0,tsfc, dsfc,bc,L, t_int)
Setup ODEFunction for LC+Bf with no Jacobian

TMItransient.stability_check — Method

stability_check(sol_array, Csfc) 
checks stability of ODE output

Arguments

sol_array: solution array

Csfc: boundary condition

Output

prints true or false

TMItransient.varying! — Method

function varying!(du, u, p, t)
ODE function for varying boundary cond
Sets up dc/dt = L*C + B*f to be solved

Arguments

du: dc/dt (must have this name for DifferentialEquations.jl to work

u: C, what we are solving for

p: parameters for diffeq - must hold specified vars

t: time we are solving for (automatically determined by DE.jl)

Output

du: numerical value of LC+Bf, vector of size 74064 for 4°